+++ docs/.vuepress/theme/layouts/Layout.vue
@@ -1,35 +1,27 @@
 <template>
   <div
-      class="theme-container"
-      :class="pageClasses"
-      @touchstart="onTouchStart"
-      @touchend="onTouchEnd"
+    class="theme-container"
+    :class="pageClasses"
+    @touchstart="onTouchStart"
+    @touchend="onTouchEnd"
   >
     <!-- Google Tag Manager (noscript) -->
     <noscript>
       <iframe
-          :src="`https://www.googletagmanager.com/ns.html?id=${ $site.themeConfig.GTM_TAG }`"
-          height="0"
-          width="0"
-          style="display:none;visibility:hidden">
+        :src="`https://www.googletagmanager.com/ns.html?id=${$site.themeConfig.GTM_TAG}`"
+        height="0"
+        width="0"
+        style="display: none; visibility: hidden"
+      >
       </iframe>
     </noscript>
     <!-- End Google Tag Manager (noscript) -->
 
-    <Navbar
-        v-if="shouldShowNavbar"
-        @toggle-sidebar="toggleSidebar"
-    />
+    <Navbar v-if="shouldShowNavbar" @toggle-sidebar="toggleSidebar" />
 
-    <div
-        class="sidebar-mask"
-        @click="toggleSidebar(false)"
-    />
+    <div class="sidebar-mask" @click="toggleSidebar(false)" />
 
-    <Sidebar
-        :items="sidebarItems"
-        @toggle-sidebar="toggleSidebar"
-    >
+    <Sidebar :items="sidebarItems" @toggle-sidebar="toggleSidebar">
       <template #top>
         <slot name="sidebar-top" />
       </template>
@@ -40,10 +32,7 @@
 
     <Home v-if="$page.frontmatter.home" />
 
-    <Page
-        v-else
-        :sidebar-items="sidebarItems"
-    >
+    <Page v-else :sidebar-items="sidebarItems">
       <template #top>
         <slot name="page-top" />
       </template>
@@ -55,97 +44,95 @@
 </template>
 
 <script>
-import Home from '@theme/components/Home.vue'
-import Navbar from '@theme/components/Navbar.vue'
-import Page from '@theme/components/Page.vue'
-import Sidebar from '@theme/components/Sidebar.vue'
-import { resolveSidebarItems } from '@vuepress/theme-default/util'
+import Home from "@theme/components/Home.vue";
+import Navbar from "@theme/components/Navbar.vue";
+import Page from "@theme/components/Page.vue";
+import Sidebar from "@theme/components/Sidebar.vue";
+import { resolveSidebarItems } from "@vuepress/theme-default/util";
 export default {
-  name: 'Layout',
+  name: "Layout",
   components: {
     Home,
     Page,
     Sidebar,
-    Navbar
+    Navbar,
   },
-  data () {
+  data() {
     return {
-      isSidebarOpen: false
-    }
+      isSidebarOpen: false,
+    };
   },
   computed: {
-    shouldShowNavbar () {
-      const { themeConfig } = this.$site
-      const { frontmatter } = this.$page
-      if (
-          frontmatter.navbar === false
-          || themeConfig.navbar === false) {
-        return false
+    shouldShowNavbar() {
+      const { themeConfig } = this.$site;
+      const { frontmatter } = this.$page;
+      if (frontmatter.navbar === false || themeConfig.navbar === false) {
+        return false;
       }
       return (
-          this.$title
-          || themeConfig.logo
-          || themeConfig.repo
-          || themeConfig.nav
-          || this.$themeLocaleConfig.nav
-      )
+        this.$title ||
+        themeConfig.logo ||
+        themeConfig.repo ||
+        themeConfig.nav ||
+        this.$themeLocaleConfig.nav
+      );
     },
-    shouldShowSidebar () {
-      const { frontmatter } = this.$page
+    shouldShowSidebar() {
+      const { frontmatter } = this.$page;
       return (
-          !frontmatter.home
-          && frontmatter.sidebar !== false
-          && this.sidebarItems.length
-      )
+        !frontmatter.home &&
+        frontmatter.sidebar !== false &&
+        this.sidebarItems.length
+      );
     },
-    sidebarItems () {
+    sidebarItems() {
       return resolveSidebarItems(
-          this.$page,
-          this.$page.regularPath,
-          this.$site,
-          this.$localePath
-      )
+        this.$page,
+        this.$page.regularPath,
+        this.$site,
+        this.$localePath
+      );
     },
-    pageClasses () {
-      const userPageClass = this.$page.frontmatter.pageClass
+    pageClasses() {
+      const userPageClass = this.$page.frontmatter.pageClass;
       return [
         {
-          'no-navbar': !this.shouldShowNavbar,
-          'sidebar-open': this.isSidebarOpen,
-          'no-sidebar': !this.shouldShowSidebar
+          "no-navbar": !this.shouldShowNavbar,
+          "sidebar-open": this.isSidebarOpen,
+          "no-sidebar": !this.shouldShowSidebar,
         },
-        userPageClass
-      ]
-    }
+        userPageClass,
+      ];
+    },
   },
-  mounted () {
+  mounted() {
     this.$router.afterEach(() => {
-      this.isSidebarOpen = false
-    })
+      this.isSidebarOpen = false;
+    });
   },
   methods: {
-    toggleSidebar (to) {
-      this.isSidebarOpen = typeof to === 'boolean' ? to : !this.isSidebarOpen
-      this.$emit('toggle-sidebar', this.isSidebarOpen)
+    toggleSidebar(to) {
+      this.isSidebarOpen = typeof to === "boolean" ? to : !this.isSidebarOpen;
+      this.$emit("toggle-sidebar", this.isSidebarOpen);
     },
     // side swipe
-    onTouchStart (e) {
+    onTouchStart(e) {
       this.touchStart = {
         x: e.changedTouches[0].clientX,
-        y: e.changedTouches[0].clientY
-      }
+        y: e.changedTouches[0].clientY,
+      };
     },
-    onTouchEnd (e) {
-      const dx = e.changedTouches[0].clientX - this.touchStart.x
-      const dy = e.changedTouches[0].clientY - this.touchStart.y
+    onTouchEnd(e) {
+      const dx = e.changedTouches[0].clientX - this.touchStart.x;
+      const dy = e.changedTouches[0].clientY - this.touchStart.y;
       if (Math.abs(dx) > Math.abs(dy) && Math.abs(dx) > 40) {
         if (dx > 0 && this.touchStart.x <= 80) {
-          this.toggleSidebar(true)
+          this.toggleSidebar(true);
         } else {
-          this.toggleSidebar(false)
+          this.toggleSidebar(false);
         }
       }
-    }
-  }
-}
+    },
+  },
+};
 </script>