+++ packages/composables/__tests__/useCustomerAddresses.spec.ts
@@ -13,11 +13,10 @@ const mockedComposables = Composables as jest.Mocked<typeof Composables>;
 const consoleErrorSpy = jest.spyOn(console, "error");
 
 import { useCustomerAddresses } from "../src/hooks/useCustomerAddresses";
+import { prepareRootContextMock } from "./contextRunner";
 describe("Composables - useCustomerAddresses", () => {
   const stateUser: Ref<Object | null> = ref(null);
-  const rootContextMock: any = {
-    $shopwareApiInstance: jest.fn(),
-  };
+  const rootContextMock = prepareRootContextMock();
 
   beforeEach(() => {
     jest.resetAllMocks();
@@ -38,6 +37,12 @@ describe("Composables - useCustomerAddresses", () => {
         refreshUser: () => {},
       } as any;
     });
+
+    mockedComposables.useVueContext.mockReturnValue({
+      isVueComponent: false,
+      isVueScope: true,
+    });
+    mockedComposables.getApplicationContext.mockReturnValue(rootContextMock);
 
     consoleErrorSpy.mockImplementationOnce(() => {});
   });
@@ -141,8 +146,8 @@ describe("Composables - useCustomerAddresses", () => {
           useCustomerAddresses(rootContextMock);
         await loadAddresses();
         expect(mockedApiClient.getCustomerAddresses).toBeCalledTimes(1);
-        expect(errors.loadAddresses).toStrictEqual([]);
-        expect(addresses.value).toBeUndefined();
+        expect(errors.loadAddresses).toBeUndefined();
+        expect(addresses.value).toStrictEqual([]);
       });
 
       it("should invoke client getCustomerAddresses method and assign error message if client request is rejected", async () => {