+++ packages/composables/src/hooks/useCustomerOrders.ts
@@ -2,8 +2,11 @@ import { ref, Ref, UnwrapRef, reactive } from "vue-demi";
 import { getCustomerOrders } from "@shopware-pwa/shopware-6-client";
 import { Order } from "@shopware-pwa/commons/interfaces/models/checkout/order/Order";
 import { ShopwareError } from "@shopware-pwa/commons/interfaces/errors/ApiError";
-import { useDefaults } from "@shopware-pwa/composables";
-import { ApplicationVueContext, getApplicationContext } from "../appContext";
+import {
+  useDefaults,
+  ApplicationVueContext,
+  getApplicationContext,
+} from "@shopware-pwa/composables";
 import { ShopwareSearchParams } from "@shopware-pwa/commons/interfaces/search/SearchCriteria";
 
 /**