+++ packages/nuxt-module/plugins/i18n.js
@@ -1,26 +1,33 @@
-import { useSharedState } from "@shopware-pwa/composables";
+import { useSharedState, extendScopeContext } from "@shopware-pwa/composables";
 import Vue from "vue";
 import VueI18n from "vue-i18n";
+import { effectScope } from "vue-demi";
 
 Vue.use(VueI18n);
 
-export default ({ app }, inject) => {
-  const { sharedRef } = useSharedState(app);
-  const currentDomainData = sharedRef("sw-current-domain");
-  // Set i18n instance on app
-  // This way we can use it in middleware and pages asyncData/fetch
-  const i18n = new VueI18n({
-    locale:
-      currentDomainData.value?.languageLocaleCode ||
-      "<%= options.defaultLanguage %>",
-    fallbackLocale: "<%= options.defaultLanguage %>",
-    messages: {
-      // <% options.availableLocales.forEach(function(availableLocale) { %>
-      "<%= availableLocale %>": require("sw-plugins/locales/<%= availableLocale %>.json"),
-      // <% }) %>
-    },
-  });
+export default async ({ app }, inject) => {
+  const scope = effectScope();
+  extendScopeContext(scope, app);
 
-  app.i18n = i18n;
-  inject("i18n", i18n);
+  await scope.run(async () => {
+    const { sharedRef } = useSharedState();
+    const currentDomainData = sharedRef("sw-current-domain");
+    // Set i18n instance on app
+    // This way we can use it in middleware and pages asyncData/fetch
+    const i18n = new VueI18n({
+      locale:
+        currentDomainData.value?.languageLocaleCode ||
+        "<%= options.defaultLanguage %>",
+      fallbackLocale: "<%= options.defaultLanguage %>",
+      messages: {
+        // <% options.availableLocales.forEach(function(availableLocale) { %>
+        "<%= availableLocale %>": require("sw-plugins/locales/<%= availableLocale %>.json"),
+        // <% }) %>
+      },
+    });
+
+    app.i18n = i18n;
+    inject("i18n", i18n);
+  });
+  scope.stop();
 };