+++ packages/nuxt-module/src/module.ts
@@ -46,6 +46,14 @@ export async function runModule(
       "@vue/composition-api/dist/vue-composition-api.esm.js"
     );
 
+  moduleObject.options.alias["vue"] =
+    moduleObject.options.alias["vue"] ||
+    (moduleObject.options.dev
+      ? moduleObject.nuxt.resolver.resolveModule("vue/dist/vue.common.dev.js")
+      : moduleObject.nuxt.resolver.resolveModule(
+          "vue/dist/vue.runtime.esm.js"
+        ));
+
   const TARGET_SOURCE: string = getTargetSourcePath(moduleObject);
   const THEME_SOURCE: string = getThemeSourcePath(
     moduleObject,
@@ -139,8 +147,10 @@ export async function runModule(
 
   let config = merge({}, getDefaultApiParams(), shopwarePwaConfig.apiDefaults);
   try {
+    /* istanbul ignore next */
     const defaultsConfigBuilder =
       require("@shopware-pwa/nuxt-module/api-defaults").default;
+    /* istanbul ignore next */
     config = defaultsConfigBuilder().get();
   } catch (e) {
     console.error("Cannot resolve API defaults config", e);